Bugfix
- Fixed issue with MSB properties that were given <x>Boolean attributes
- Fixed an issue with the Model Editor that blocked editing properties

+++ src/StudioCore/Editors/MapEditor/MapPropertyEditor.cs
@@ -159,7 +159,7 @@ public class MapPropertyEditor
                     else
                         val = 0;
 
-                    newval = val;
+                    newval = (short)val;
                     isChanged = true;
                 }
             }
@@ -229,7 +229,7 @@ public class MapPropertyEditor
                     else
                         val = 0;
 
-                    newval = val;
+                    newval = (sbyte)val;
                     isChanged = true;
                 }
             }
